usbredirparser: implement log callback, backed to log crate
Implement the log callback, passing the actual level and message to our
hook function.

However, rather than calling into a log callback on ParserHandler, use
the log crate directly. Since a handler may want to call into the parser
as part of its own callbacks (eg, to reply to a packet), the parser C
implementation is likely to call log(), which would then call *back*
into the handler. This would violate our borrow rules (we would need two
mutable references to the handler).

Since we have a common log framework in Rust, just use that for the
logging instead.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>