usbredirparser: pass parser into handler callbacks
Currently, a ParserHandler impl can't do much: it cannot embed a
reference to the Parser itself (as that would be cyclic; the Parser owns
the handler), and so cannot call into the parser - for example, to
respond to an incoming transfer.

Change the current callbacks to pass a (non-mutable) parser reference as
a parameter to each. This allows the handler to send USB responses
through the parser.

To allow this, we change the C parser's priv pointer to reference the
parser, allowing access to both the Rust Parser and Handler structs.
Since we're using the pointer directly, we need to Pin the parser (we
had not done so with the handler previously).

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>